[Common] id in aggregation
Fix index value type in tensor-aggregation.

+++ gst/nnstreamer/elements/gsttensor_aggregator.c
@@ -521,7 +521,7 @@ static GstAdapter *
 gst_tensor_aggregator_get_adapter (GstTensorAggregator * self, GstBuffer * buf)
 {
   GstMetaQuery *meta;
-  guint32 key = 0;
+  gint64 key = 0;
 
   meta = gst_buffer_get_meta_query (buf);
   if (meta)
